chg: addressing a few comments left from previous pr #19390 #19392

+++ src/app/clusters/level-control/level-control.cpp
@@ -141,12 +141,12 @@ static void timerCallback(System::Layer *, void * callbackContext)
 static void schedule(EndpointId endpoint, uint32_t delayMs)
 {
     DeviceLayer::SystemLayer().StartTimer(chip::System::Clock::Milliseconds32(delayMs), timerCallback,
-                                          reinterpret_cast<void *>(endpoint));
+                                          reinterpret_cast<void *>(static_cast<uintptr_t>(endpoint)));
 }
 
 static void deactivate(EndpointId endpoint)
 {
-    DeviceLayer::SystemLayer().CancelTimer(timerCallback, reinterpret_cast<void *>(endpoint));
+    DeviceLayer::SystemLayer().CancelTimer(timerCallback, reinterpret_cast<void *>(static_cast<uintptr_t>(endpoint)));
 }
 
 static EmberAfLevelControlState * getState(EndpointId endpoint)
